+++ src/test/java/eu/nicosworld/rithmo/core/turn/resolver/CaptureResolverTest.java
@@ -1,7 +1,13 @@
 package eu.nicosworld.rithmo.core.turn.resolver;
 
+import eu.nicosworld.rithmo.core.GameOptions;
+import eu.nicosworld.rithmo.core.game.CaptureRuleOption;
+import eu.nicosworld.rithmo.core.game.VictoryRuleOption;
 import eu.nicosworld.rithmo.core.helper.TestDebugger;
+import eu.nicosworld.rithmo.core.turn.TurnPhase;
+import eu.nicosworld.rithmo.core.turn.TurnState;
 import eu.nicosworld.rithmo.core.turn.option.PreCaptureOption;
+import eu.nicosworld.rithmo.engine.capture.capturerule.PowerRule;
 import eu.nicosworld.rithmo.engine.capture.model.CaptureAction;
 import eu.nicosworld.rithmo.engine.capture.model.CaptureContext;
 import eu.nicosworld.rithmo.engine.capture.CaptureEngine;
@@ -17,6 +23,8 @@ import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
 
 import java.util.List;
+import java.util.Map;
+import java.util.Set;
 
 import static org.junit.jupiter.api.Assertions.*;
 
@@ -298,8 +306,28 @@ class CaptureResolverTest {
         }
     }
 
-    @Test
-    void resolvePostCaptures_WithPyramid() {
+    @Nested
+    class TwoRulesEngine_EncounterAndPower {
+        @BeforeEach
+        void setup() {
+            EncounterRule encounterRule = new EncounterRule(regularMoveGenerator, freePathMovementValidator);
+            PowerRule powerRule = new PowerRule(regularMoveGenerator, freePathMovementValidator);
+
+            captureEngine = new CaptureEngine(List.of(encounterRule, powerRule));
+            captureResolver = new CaptureResolver(captureEngine);
+        }
+
+        @Test
+        void resolvePostCaptures_WithPyramids_EncounterAndPower() {
+            Board board = new BoardBuilder(4, 4)
+                    .fullBlackPyramidAt(0, 0)
+                    .fullWhitePyramidAt(0, 3)
+                    .build();
+
+            List<PreCaptureOption> choices = captureResolver.resolvePreCaptures(GameState.initial(board, black));
+
+            assertEquals(6, choices.size());
+        }
     }